Added initial revision of pg_stat_statements normalization. Tests are included, written in Python, with a dependency on psycopg2.
Please note the following caveats - my initial focus was on proving the
concept, so I neglected to do a certain amount of plumbing, preferring
to focus on test cases and the general concept.

There are currently two compiler warnings due to the loss of volatility
of a pointed-to struct at one point. This is a concern, so I haven't
suppressed the warnings artificially. I'll revisit this in a while.

Some pg_stat_statements functionality has been gutted, but will be
restored in a later revision. In particular, serialization to disk was
removed.

The last test currently fails. I'm working on this.

The current implementation is perfectly happy to write past the end of
an array that it writes it's selective serialisation to for the current
query, if it encounters a sufficiently long query.

I'm planning on replacing this with a circular buffer
abstraction, that hashes the last iteration, appends the hash to the
next iteration, and makes as many iterations as it has to, though often
only one. The hash function I'm using displays the avalanche effect, so
this shouldn't be a problem (it's not a cryptographic hash function or
anything silly like that - it's the same function that was previously
used to hash the query string).