Deal with memory reordering issues when manipulating spinlock protected data
On reflection, there doesn't seem to be an obviously superior alternative
available to spinlocks for protecting counters/query string for updating.

While I have my doubts that the number of instructions executed while the
spinlock is held is just "a few" (as comments on the spinlock macros suggest
that number be limited to), it seems prudent to use spinlocks correctly for
as long as we're using them.

In passing, add another test.