Defer hot standby initialization until minimum recovery point
Previously, the hot standby initialization could occur before enough
WAL has been replayed to bring the heap to a consistent state -- the
minimum recovery point.  This happens to satisfy an invariant that
postmaster will not be told to accept connections until those steps
have been performed.  Maintain that invariant, but only go through
with initialization should a minimum recovery point be reached first;
previously the initialization code would be called on a data directory
that was not yet consistent, and would attempt to read the status of
transactions not yet replayed via WAL.

Given that the initialization code may be called many times, guard it
to prevent multiple initialization.