Fix issue with Const canonicalization.
With this commit, even if we run out of space in the jumble we'll
continue to walk the query tree to find Const nodes to canonicalize.

It's always possible for a query string to be byte-for-byte larger
than the jumble representation, so this is important. However, we give
up having exceeded JUMBLE_SIZE twice over, to prevent the stack
overflows that might otherwise occur, particularly in the face of
ridiculously heavily nested set operations.