Fix utility statements + prepared statements. Support legacy behavior.
We now directly hash the query string under some circumstances. It
wouldn't have been worth normalizing utility statements such that minor
differences don't differentiate. The query string that the user sees and
that's hashed is constant across runs of the same parameterized query,
so it should be perfectly stable. It won't be recognised as equivalent
if the query is re-prepared and it has some difference that would not
generally be considered a differentiator, but that doesn't seem worth
fixing.

Support legacy query string based matching in case somebody actually
misses that behavior.

Minor re-wording of pg_stat_statements.c header.