Remove Jumble parameter. Make legacy behavior guc PGC_POSTMASTER.
There was no point in passing around a pointer to last_jumble - it's
static global data, as are some other variables that store information
about the query tree, so let's not pretend otherwise.

Users shouldn't be allowed to march to the beat of their own drum by
setting legacy behavior locally, even superusers.

Also, further polishing of comments, for clarity and consistency with
base code.