Fix problems with planner calls in nested queries. Further refactoring.
The implementation of pg_stat_statements jumble-based query
normalization was previously liable to get confused in the face of
nested executor calls, such as SPI query execution, as this sometimes
resulted in the queries being planned and pgss_Planner overwriting the
useful jumble with one for a query that would often never be recorded.

The implementation should now be well-behaved, even in the face of
queries calling functions that execute queries with
pg_stat_statements.track set to 'all'. We now only jumble a completely
unnested query, regardless of the 'track' setting.

To ensure that it is well-behaved, relevant tests were added.

Abandon scheme for avoiding walking the entire query tree. This was
probably unnecessary, and certainly ugly.

This might need some more polishing before a further submission to
-hackers, but it's certainly fairly close.